// Frontend/src/api/models/ProjectUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Project update schema
 */
export type ProjectUpdate = {
    manager_id?: (number | null);
    end_date?: (string | null);
    objective?: (string | null);
    status?: (string | null);
    planned_cost?: (number | string | null);
    actual_cost?: (number | string | null);
    planned_effort?: (number | string | null);
    actual_effort?: (number | string | null);
    baseline_start?: (string | null);
    baseline_finish?: (string | null);
    progress?: (number | null);
};